IS: Refactor services into application folder
