+++ src/main/kotlin/no/nav/syfo/aktivitetskravvurdering/kafka/AktivitetskravVurderingConsumer.kt
@@ -1,6 +1,6 @@
 package no.nav.syfo.aktivitetskravvurdering.kafka
 
-import no.nav.syfo.personstatus.PersonoversiktStatusService
+import no.nav.syfo.personstatus.application.PersonoversiktStatusService
 import no.nav.syfo.personstatus.domain.PersonIdent
 import no.nav.syfo.personstatus.infrastructure.kafka.KafkaConsumerService
 import org.apache.kafka.clients.consumer.ConsumerRecords
