+++ src/main/kotlin/no/nav/syfo/aktivitetskravvurdering/kafka/KafkaAktivitetskravVurderingTask.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.aktivitetskravvurdering.kafka
 
 import no.nav.syfo.ApplicationState
-import no.nav.syfo.personstatus.PersonoversiktStatusService
+import no.nav.syfo.personstatus.application.PersonoversiktStatusService
 import no.nav.syfo.personstatus.infrastructure.kafka.KafkaEnvironment
 import no.nav.syfo.personstatus.infrastructure.kafka.kafkaAivenConsumerConfig
 import no.nav.syfo.personstatus.infrastructure.kafka.launchKafkaTask
